chore: filtering remade in a simply way
