+++ src/redux/hooks.ts
@@ -1,7 +1,7 @@
 import { TypedUseSelectorHook, useDispatch, useSelector } from "react-redux";
 import { ThunkDispatch } from "redux-thunk";
-import { ProjectActions } from "./actions";
 import { RootState } from "./store";
+import { AnyAction } from "redux";
 
-export const useAppDispatch = useDispatch<ThunkDispatch<RootState,void,ProjectActions>>
+export const useAppDispatch = useDispatch<ThunkDispatch<RootState,void,AnyAction>>
 export const useAppSelector :TypedUseSelectorHook<RootState> = useSelector
