--- urls
@@ -1,5 +1,0 @@
-contacts: 
-https://mocki.io/v1/9537a935-a0d9-48e3-806c-b5de95b8ce35
-
-groups: 
-https://mocki.io/v1/0bad626f-bb05-42e7-9915-baaf710f8ba5
